refactor(api): split recommender selection modules
